+++ src/main/java/com/imooc/repository/HouseRepository.java
@@ -39,6 +39,6 @@ import com.imooc.entity.House;
  * @Copyright: 2019 www.xxx.com Inc. All rights reserved. 
  * 注意：本内容仅限于公司内部传阅，禁止外泄以及用于其他的商业目 
  */
-public interface HouseRepository extends PagingAndSortingRepository<House, Long>, JpaSpecificationExecutor<House> {
+public interface HouseRepository extends PagingAndSortingRepository<House, Integer>, JpaSpecificationExecutor<House> {
 
 }
