+++ src/main/java/com/imooc/web/form/HouseForm.java
@@ -45,7 +45,7 @@ import com.alibaba.fastjson.JSON;
  */
 public class HouseForm {
 
-	private Long id;
+	private int id;
 
 	@NotNull(message = "大标题不允许为空")
 	@Size(min = 1, max = 30, message = "标题长度必须在1-30之间")
@@ -103,8 +103,8 @@ public class HouseForm {
 	@Max(value=1)
 	private Integer rentWay;
 	
-	private Long subwayLineId;
-	private Long subwayStationId;
+	private int subwayLineId;
+	private int subwayStationId;
 	private int distanceToSubway = -1;
 	private String layoutDesc;
 	private String roundService;
@@ -116,11 +116,11 @@ public class HouseForm {
 	private List<String> tags;
 	private List<PhotoForm> photos;
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
@@ -244,19 +244,19 @@ public class HouseForm {
 		this.rentWay = rentWay;
 	}
 
-	public Long getSubwayLineId() {
+	public int getSubwayLineId() {
 		return subwayLineId;
 	}
 
-	public void setSubwayLineId(Long subwayLineId) {
+	public void setSubwayLineId(int subwayLineId) {
 		this.subwayLineId = subwayLineId;
 	}
 
-	public Long getSubwayStationId() {
+	public int getSubwayStationId() {
 		return subwayStationId;
 	}
 
-	public void setSubwayStationId(Long subwayStationId) {
+	public void setSubwayStationId(int subwayStationId) {
 		this.subwayStationId = subwayStationId;
 	}
